# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-Native Data Platform API",
    "description": "Governed local data platform: authentication and RBAC, multi-engine query gateway, DataHub-backed catalog and lineage, versioned metrics, and checkpointed evidence-bound analyses.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/datasources": {
      "get": {
        "tags": [
          "datasources"
        ],
        "summary": "List Datasources",
        "operationId": "list_datasources_api_v1_datasources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasourceListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "datasources"
        ],
        "summary": "Create Datasource",
        "operationId": "create_datasource_api_v1_datasources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasourceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/queries": {
      "post": {
        "tags": [
          "queries"
        ],
        "summary": "Submit Query",
        "operationId": "submit_query_api_v1_queries_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuerySubmitRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuerySubmitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "queries"
        ],
        "summary": "List Queries",
        "operationId": "list_queries_api_v1_queries_get",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/queries/{query_id}": {
      "get": {
        "tags": [
          "queries"
        ],
        "summary": "Get Query",
        "operationId": "get_query_api_v1_queries__query_id__get",
        "parameters": [
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Query Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/queries/{query_id}/results": {
      "get": {
        "tags": [
          "queries"
        ],
        "summary": "Get Results",
        "operationId": "get_results_api_v1_queries__query_id__results_get",
        "parameters": [
          {
            "name": "query_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Query Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/roles": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Roles",
        "operationId": "list_roles_api_v1_admin_roles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleResponse"
                  },
                  "type": "array",
                  "title": "Response List Roles Api V1 Admin Roles Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Role",
        "operationId": "create_role_api_v1_admin_roles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/grants": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Grants",
        "operationId": "list_grants_api_v1_admin_grants_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Grant",
        "operationId": "create_grant_api_v1_admin_grants_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrantResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DatasourceCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._-]+$",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "enum": [
              "postgres",
              "mysql",
              "doris"
            ],
            "title": "Kind"
          },
          "connection_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PostgresConnectionConfig"
              },
              {
                "$ref": "#/components/schemas/MySQLConnectionConfig"
              },
              {
                "$ref": "#/components/schemas/DorisConnectionConfig"
              }
            ],
            "title": "Connection Config"
          },
          "secret_ref": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]*$",
            "title": "Secret Ref"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "kind",
          "connection_config",
          "secret_ref"
        ],
        "title": "DatasourceCreate"
      },
      "DatasourceListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DatasourceResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "DatasourceListResponse"
      },
...
      "DatasourceResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "connection_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Connection Config"
          },
          "secret_ref": {
            "type": "string",
            "title": "Secret Ref"
          },
          "capabilities": {
            "additionalProperties": true,
            "type": "object",
            "title": "Capabilities"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "health_status": {
            "type": "string",
            "title": "Health Status"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind",
          "connection_config",
          "secret_ref",
          "capabilities",
          "enabled",
          "health_status",
          "created_by",
          "created_at",
          "version"
        ],
        "title": "DatasourceResponse"
      },
...
      "GrantListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GrantResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "GrantListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/ProfileResponse"
          },
          "csrf_token": {
            "type": "string",
            "title": "Csrf Token"
          }
        },
        "type": "object",
        "required": [
          "profile",
          "csrf_token"
        ],
        "title": "LoginResponse"
      },
...
      "ProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "roles",
          "capabilities"
        ],
        "title": "ProfileResponse"
      },
...
      "QueryDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "datasource_id": {
            "type": "string",
            "format": "uuid",
            "title": "Datasource Id"
          },
          "analysis_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "sql": {
            "type": "string",
            "title": "Sql"
          },
          "validated_sql": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validated Sql"
          },
          "parameters": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parameters"
          },
          "sql_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sql Hash"
          },
          "limits": {
            "additionalProperties": true,
            "type": "object",
            "title": "Limits"
          },
          "policy_revision": {
            "type": "integer",
            "title": "Policy Revision"
          },
          "attempt": {
            "type": "integer",
            "title": "Attempt"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorBody"
              },
              {
                "type": "null"
              }
            ]
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResultSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "cancel_requested_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancel Requested At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "datasource_id",
          "status",
          "purpose",
          "sql",
          "policy_revision",
          "attempt",
          "created_at"
        ],
        "title": "QueryDetail"
      },
...
      "QueryListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QueryDetail"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "QueryListResponse"
      },
      "QueryResultResponse": {
        "properties": {
          "query_id": {
            "type": "string",
            "format": "uuid",
            "title": "Query Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResultPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          }
        },
        "type": "object",
        "required": [
          "query_id",
          "status",
          "trace_id"
        ],
        "title": "QueryResultResponse"
      },
      "QuerySubmitRequest": {
        "properties": {
          "datasource_id": {
            "type": "string",
            "format": "uuid",
            "title": "Datasource Id"
          },
          "sql": {
            "type": "string",
            "minLength": 1,
            "title": "Sql"
          },
          "parameters": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue-Input"
            },
            "type": "object",
            "title": "Parameters"
          },
          "limits": {
            "$ref": "#/components/schemas/QueryLimits"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "query",
              "explain"
            ],
            "title": "Purpose",
            "default": "query"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "datasource_id",
          "sql"
        ],
        "title": "QuerySubmitRequest"
      },
      "QuerySubmitResponse": {
        "properties": {
          "query_id": {
            "type": "string",
            "format": "uuid",
            "title": "Query Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "query_id",
          "status"
        ],
        "title": "QuerySubmitResponse"
      },
...
      "RoleCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-z][a-z0-9_-]*$",
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "RoleCreate"
      },
      "RoleResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "capabilities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Capabilities"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "capabilities"
        ],
        "title": "RoleResponse"
      },
...
    }
  },
...
}